requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
piexif==1.1.3
numpy==1.26.3
jinja2==3.1.3
aiofiles==23.2.1
requests==2.31.0
